requirements.txt
Django==4.1.6
djangorestframework==3.14.0
psycopg2-binary==2.9.5
drf-yasg==1.21.5
django-filter~=22.1
pytest==7.2.1
pytest-django==4.5.2
django-phonenumber-field==7.0.2
PyJWT~=2.6.0